Struct hexasphere::shapes::IcoSphereBase

pub struct IcoSphereBase;

Implements an icosahedron as the base shape.

12 vertices

20 faces

30 edges

This shape has the best results for a sphere.

The resulting smaller triangles are close to being equilateral, so if one draws lines from the center of the each triangle to the middle of the each edge then the result will be 12 pentagons and many hexagons.