Trait hexasphere::EquilateralBaseShape

pub trait EquilateralBaseShape: BaseShape {
    fn triangle_normals() -> &'static [Vec3A];

    fn triangle_min_dot() -> f32;
}

Implemented in the case where the triangles on the shape are both equilateral and identifiable from their normal.

This is only used in the cases of spherical shapes.

Required methods

fn triangle_normals() -> &'static [Vec3A]

Normals for each of the triangles provided by BaseShape::triangles.

fn triangle_min_dot() -> f32

Minimum value for the dot product which one could use to determine that triangle being the closest.

If the dot product between a vector and a triangle normal is greater than this, then that vector is guaranteed to be within that triangle.

This is also the cosine of the angle of the cone whose circular edge lies on a triangle.

Loading content...